+++ rspice/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+
+- Improve ruby 3.0 compatibility in `a_class_pass_method` shared example ([#440](https://github.com/Freshly/spicerack/pull/440))
+
 ## v0.25.6
 
 *Release Date*: 2/24/2021